refactor: rearrange types in controls pkg
