--- packages/controls/src/common/data-type-key.ts
@@ -1,1 +1,0 @@
-export const ControlDataTypeKey = '@@makeswift/type'
