+++ packages/controls/src/common/index.ts
@@ -1,2 +1,1 @@
-export * from './data-type-key'
 export * from './types'
