+++ packages/controls/src/index.ts
@@ -1,5 +1,3 @@
-export * from './common'
-
 export * from './checkbox'
 export * from './color'
 export * from './combobox'
@@ -9,3 +7,5 @@ export * from './link'
 export * from './number'
 export * from './text-input'
 export * from './text-area'
+
+export * from './common/types'
